fixed firebase prev list
