+++ package.json
@@ -10,6 +10,7 @@
     "jquery": "^3.3.1",
     "material-ui": "^0.20.0",
     "popper.js": "^1.14.3",
+    "re-base": "^3.2.2",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
     "react-router-dom": "^4.2.2",
