+++ src/index.js
@@ -10,17 +10,12 @@ import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 import * as firebase from 'firebase';
+import AOS  from 'aos';
+import "../node_modules/aos/dist/aos.css"
+
+AOS.init();
 
 
-var config = {
-    apiKey: "AIzaSyCnNq6ZJIy_czx-0d7viiPEZAiUA4K2XBk",
-    authDomain: "food-for-thought-59940.firebaseapp.com",
-    databaseURL: "https://food-for-thought-59940.firebaseio.com",
-    projectId: "food-for-thought-59940",
-    storageBucket: "food-for-thought-59940.appspot.com",
-    messagingSenderId: "130909874537"
-  };
-  firebase.initializeApp(config);
